Lockhammer: move types to header file
Move argument types to a header file in part to clean up main
source file.  This change also exposes the argument containing
struct to the test so that the test-specific argument parser
can utilize the values of the generic command line arguments.
